refac: moved command list functions into the type
